(Sliding Window / 2 pointers)  + Binary Search approach
ALWAYS REMEMBER SLIDING WINDOW WORKS WHEN PROPERTY IS MONOTONIC IN DIRECTION OF SLIDING, i.e. adding elements either increases/decreases and removing elements decreases/increases the property.

And same for BS obv, you throw the other half based on the MONOTONICITY, i.e the predicate is true for 1 half and false for the half that is to be thrown.